Graph classes refactor and basic cloudformation graph (#1374)
* created abstract GraphManager and LocalGraph

* moved UTs

* fixed path in test

* Created an abstract Block, BlockType, removed encoding

* removed decoding, reordered utils functions

* Created an abstract Block, BlockType, removed encoding

* fixed imports and using encode

* removed `filter_excluded_dict_attributes` and added @abstractmethod annotations

* used dict comparison
