+++ checkov/terraform/graph_builder/graph_components/block_types.py
@@ -1,13 +1,14 @@
-from enum import Enum
+from checkov.common.graph.graph_builder.graph_components.block_types import BlockType
+from dataclasses import dataclass
 
 
-class BlockType(str, Enum):
+@dataclass
+class BlockType(BlockType):
     DATA = "data"
     LOCALS = "locals"
     MODULE = "module"
     OUTPUT = "output"
     PROVIDER = "provider"
-    RESOURCE = "resource"
     TERRAFORM = "terraform"
     TF_VARIABLE = "tfvar"
     VARIABLE = "variable"
