+++ checkov/terraform/graph_builder/graph_components/module.py
@@ -4,7 +4,8 @@ from typing import List, Dict, Any, Set, Callable
 
 from checkov.terraform.checks.utils.dependency_path_handler import unify_dependency_path
 from checkov.terraform.graph_builder.graph_components.block_types import BlockType
-from checkov.terraform.graph_builder.graph_components.blocks import Block, get_inner_attributes
+from checkov.terraform.graph_builder.graph_components.blocks import TerraformBlock
+from checkov.common.graph.graph_builder.graph_components.blocks import get_inner_attributes
 
 
 class Module:
@@ -12,19 +13,17 @@ class Module:
         self,
         source_dir: str,
         module_dependency_map: Dict[str, List[List[str]]],
-        dep_index_mapping: Dict[str, str],
-        encode: bool,
+        dep_index_mapping: Dict[str, str]
     ) -> None:
         self.dep_index_mapping = dep_index_mapping
         self.module_dependency_map = module_dependency_map
         self.path = ""
-        self.blocks: List[Block] = []
+        self.blocks: List[TerraformBlock] = []
         self.customer_name = ""
         self.account_id = ""
         self.source = ""
         self.resources_types: Set[str] = set()
         self.source_dir = source_dir
-        self.encode = encode
 
     def add_blocks(
         self, block_type: BlockType, blocks: List[Dict[str, Dict[str, Any]]], path: str, source: str
@@ -33,7 +32,7 @@ class Module:
         if self._block_type_to_func.get(block_type):
             self._block_type_to_func[block_type](self, blocks, path)
 
-    def _add_to_blocks(self, block: Block) -> None:
+    def _add_to_blocks(self, block: TerraformBlock) -> None:
         dependencies = [dep_trail for dep_trail in self.module_dependency_map.get(os.path.dirname(block.path), [])]
         module_dependency_num = ""
         if not dependencies:
@@ -54,14 +53,13 @@ class Module:
                 alias = attributes.get("alias")
                 if alias:
                     provider_name = f"{provider_name}.{alias[0]}"
-                provider_block = Block(
+                provider_block = TerraformBlock(
                     block_type=BlockType.PROVIDER,
                     name=provider_name,
                     config=provider_dict,
                     path=path,
                     attributes=attributes,
-                    source=self.source,
-                    encode=self.encode,
+                    source=self.source
                 )
                 self._add_to_blocks(provider_block)
 
@@ -69,28 +67,26 @@ class Module:
         for variable_dict in blocks:
             for name in variable_dict:
                 attributes = variable_dict[name]
-                variable_block = Block(
+                variable_block = TerraformBlock(
                     block_type=BlockType.VARIABLE,
                     name=name,
                     config=variable_dict,
                     path=path,
                     attributes=attributes,
-                    source=self.source,
-                    encode=self.encode,
+                    source=self.source
                 )
                 self._add_to_blocks(variable_block)
 
     def _add_locals(self, blocks: List[Dict[str, Dict[str, Any]]], path: str) -> None:
         for blocks_section in blocks:
             for name in blocks_section:
-                local_block = Block(
+                local_block = TerraformBlock(
                     block_type=BlockType.LOCALS,
                     name=name,
                     config={name: blocks_section[name]},
                     path=path,
                     attributes={name: blocks_section[name]},
-                    source=self.source,
-                    encode=self.encode,
+                    source=self.source
                 )
                 self._add_to_blocks(local_block)
 
@@ -99,28 +95,26 @@ class Module:
             for name in output_dict:
                 if type(output_dict[name]) is not dict:
                     continue
-                output_block = Block(
+                output_block = TerraformBlock(
                     block_type=BlockType.OUTPUT,
                     name=name,
                     config=output_dict,
                     path=path,
                     attributes={"value": output_dict[name].get("value")},
-                    source=self.source,
-                    encode=self.encode,
+                    source=self.source
                 )
                 self._add_to_blocks(output_block)
 
     def _add_module(self, blocks: List[Dict[str, Dict[str, Any]]], path: str) -> None:
         for module_dict in blocks:
             for name in module_dict:
-                module_block = Block(
+                module_block = TerraformBlock(
                     block_type=BlockType.MODULE,
                     name=name,
                     config=module_dict,
                     path=path,
                     attributes=module_dict[name],
-                    source=self.source,
-                    encode=self.encode,
+                    source=self.source
                 )
                 self._add_to_blocks(module_block)
 
@@ -134,15 +128,14 @@ class Module:
                     if provisioner:
                         self._handle_provisioner(provisioner, attributes)
                     attributes["resource_type"] = [resource_type]
-                    resource_block = Block(
+                    resource_block = TerraformBlock(
                         block_type=BlockType.RESOURCE,
                         name=f"{resource_type}.{name}",
                         config=resource_dict,
                         path=path,
                         attributes=attributes,
                         id=f"{resource_type}.{name}",
-                        source=self.source,
-                        encode=self.encode,
+                        source=self.source
                     )
                     self._add_to_blocks(resource_block)
 
@@ -150,42 +143,39 @@ class Module:
         for data_dict in blocks:
             for data_type in data_dict:
                 for name in data_dict[data_type]:
-                    data_block = Block(
+                    data_block = TerraformBlock(
                         block_type=BlockType.DATA,
                         name=data_type + "." + name,
                         config=data_dict,
                         path=path,
                         attributes=data_dict.get(data_type, {}).get(name, {}),
                         id=data_type + "." + name,
-                        source=self.source,
-                        encode=self.encode,
+                        source=self.source
                     )
                     self._add_to_blocks(data_block)
 
     def _add_terraform_block(self, blocks: List[Dict[str, Dict[str, Any]]], path: str) -> None:
         for terraform_dict in blocks:
             for name in terraform_dict:
-                terraform_block = Block(
+                terraform_block = TerraformBlock(
                     block_type=BlockType.TERRAFORM,
                     name=name,
                     config=terraform_dict,
                     path=path,
                     attributes={},
-                    source=self.source,
-                    encode=self.encode,
+                    source=self.source
                 )
                 self._add_to_blocks(terraform_block)
 
     def _add_tf_var(self, blocks: Dict[str, Dict[str, Any]], path: str) -> None:
         for tf_var_name, attributes in blocks.items():
-            tfvar_block = Block(
+            tfvar_block = TerraformBlock(
                 block_type=BlockType.TF_VARIABLE,
                 name=tf_var_name,
                 config={tf_var_name: attributes},
                 path=path,
                 attributes=attributes,
-                source=self.source,
-                encode=self.encode,
+                source=self.source
             )
             self._add_to_blocks(tfvar_block)
