+++ tests/terraform/graph/db_connector/test_networkx_connector.py
@@ -2,7 +2,7 @@ import os
 from unittest import TestCase
 
 from checkov.common.graph.db_connectors.networkx.networkx_db_connector import NetworkxConnector
-from checkov.terraform.graph_builder.local_graph import LocalGraph
+from checkov.terraform.graph_builder.local_graph import TerraformLocalGraph
 from checkov.terraform.parser import Parser
 
 TEST_DIRNAME = os.path.dirname(os.path.realpath(__file__))
@@ -13,7 +13,7 @@ class TestNetworkxConnector(TestCase):
         resources_dir = os.path.realpath(os.path.join(TEST_DIRNAME, '../resources/encryption'))
         hcl_config_parser = Parser()
         module, module_dependency_map, _ = hcl_config_parser.parse_hcl_module(resources_dir, 'AWS')
-        local_graph = LocalGraph(module, module_dependency_map)
+        local_graph = TerraformLocalGraph(module, module_dependency_map)
         local_graph._create_vertices()
         nxc = NetworkxConnector()
         nxc.save_graph(local_graph)
