+++ tests/terraform/graph/graph_builder/test_graph_builder.py
@@ -5,7 +5,7 @@ from unittest import TestCase
 from checkov.common.graph.db_connectors.networkx.networkx_db_connector import NetworkxConnector
 from checkov.terraform.graph_builder.graph_components.block_types import BlockType
 from checkov.terraform.graph_builder.graph_to_tf_definitions import convert_graph_vertices_to_tf_definitions
-from checkov.terraform.graph_manager import GraphManager
+from checkov.terraform.graph_manager import TerraformGraphManager
 from checkov.terraform.parser import external_modules_download_path
 
 TEST_DIRNAME = os.path.dirname(os.path.realpath(__file__))
@@ -16,7 +16,7 @@ class TestGraphBuilder(TestCase):
     def test_build_graph(self):
         resources_dir = os.path.join(TEST_DIRNAME, '../resources/general_example')
 
-        graph_manager = GraphManager(db_connector=NetworkxConnector())
+        graph_manager = TerraformGraphManager(db_connector=NetworkxConnector())
         graph, tf_definitions = graph_manager.build_graph_from_source_directory(resources_dir)
 
         expected_num_of_var_nodes = 3
@@ -59,10 +59,10 @@ class TestGraphBuilder(TestCase):
             if graph.vertices[edge.origin].get_hash() == hashed_from and graph.vertices[edge.dest].get_hash() == hashed_to:
                 matching_edges.append(edge)
         self.assertGreater(len(matching_edges), 0,
-                           f'expected to find edge from [{node_from.block_type.value} {node_from.name}] to [{node_to.block_type.value} {node_to.name}] with label [{expected_label}]')
+                           f'expected to find edge from [{node_from.block_type} {node_from.name}] to [{node_to.block_type} {node_to.name}] with label [{expected_label}]')
         if not any(e.label == expected_label for e in matching_edges):
             self.fail(
-                f'expected to find edge from [{node_from.block_type.value} {node_from.name}] to [{node_to.block_type.value} {node_to.name}] with label [{expected_label}], found edges: {[str(e) for e in matching_edges]}')
+                f'expected to find edge from [{node_from.block_type} {node_from.name}] to [{node_to.block_type} {node_to.name}] with label [{expected_label}], found edges: {[str(e) for e in matching_edges]}')
 
     @staticmethod
     def get_vertex_by_name_and_type(local_graph, block_type, name, multiple=False):
@@ -73,7 +73,7 @@ class TestGraphBuilder(TestCase):
 
     def test_update_vertices_configs_deep_nesting(self):
         resources_dir = os.path.join(TEST_DIRNAME, '../resources/variable_rendering/render_deep_nesting')
-        graph_manager = GraphManager(NetworkxConnector())
+        graph_manager = TerraformGraphManager(NetworkxConnector())
         local_graph, _ = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=True)
         expected_config = {'aws_s3_bucket': {'default': {'server_side_encryption_configuration': [
             {'rule': [{'apply_server_side_encryption_by_default': [
@@ -86,7 +86,7 @@ class TestGraphBuilder(TestCase):
         # see the image to view the expected graph in tests/resources/modules/linked_modules/expected_graph.png
         resources_dir = os.path.realpath(os.path.join(TEST_DIRNAME, '../resources/modules/linked_modules'))
 
-        graph_manager = GraphManager(NetworkxConnector())
+        graph_manager = TerraformGraphManager(NetworkxConnector())
         local_graph, tf_definitions = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=False)
 
         vertices_by_block_type = local_graph.vertices_by_block_type
@@ -126,7 +126,7 @@ class TestGraphBuilder(TestCase):
         resources_dir = os.path.realpath(
             os.path.join(TEST_DIRNAME, '../resources/modules/registry_security_group_inner_module'))
 
-        graph_manager = GraphManager(NetworkxConnector())
+        graph_manager = TerraformGraphManager(NetworkxConnector())
         local_graph, tf_definitions = graph_manager.build_graph_from_source_directory(resources_dir,
                                                                                       render_variables=True,
                                                                                       download_external_modules=True)
@@ -157,7 +157,7 @@ class TestGraphBuilder(TestCase):
     def test_build_graph_with_deep_nested_edges(self):
         resources_dir = os.path.realpath(os.path.join(TEST_DIRNAME, '../resources/k8_service'))
 
-        graph_manager = GraphManager(NetworkxConnector())
+        graph_manager = TerraformGraphManager(NetworkxConnector())
         local_graph, tf = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=True)
 
         resource_kubernetes_deployment = self.get_vertex_by_name_and_type(local_graph, BlockType.RESOURCE,
@@ -178,7 +178,7 @@ class TestGraphBuilder(TestCase):
 
     def test_blocks_from_local_graph_module(self):
         resources_dir = os.path.realpath(os.path.join(TEST_DIRNAME, '../resources/modules/stacks'))
-        graph_manager = GraphManager(NetworkxConnector())
+        graph_manager = TerraformGraphManager(NetworkxConnector())
         local_graph, tf = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=True)
         tf, _ = convert_graph_vertices_to_tf_definitions(local_graph.vertices, resources_dir)
         found_results = 0
