+++ tests/terraform/graph/variable_rendering/test_renderer.py
@@ -2,6 +2,7 @@ import os
 from unittest.case import TestCase
 
 from checkov.terraform.graph_builder.graph_components.block_types import BlockType
+from checkov.terraform.graph_manager import TerraformGraphManager
 from checkov.terraform.graph_manager import GraphManager
 from tests.terraform.graph.variable_rendering.expected_data import *
 
@@ -16,58 +17,58 @@ class TestRenderer(TestCase):
 
     def test_render_local(self):
         resources_dir = os.path.join(TEST_DIRNAME, '../resources/variable_rendering/render_local')
-        graph_manager = GraphManager('acme', ['acme'])
+        graph_manager = TerraformGraphManager('acme', ['acme'])
         local_graph, _ = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=True)
 
         expected_local = {'bucket_name': 'test_bucket_name'}
         expected_resource = {'region': 'us-west-2', 'bucket': expected_local['bucket_name']}
 
-        self.compare_vertex_attributes(local_graph, expected_local, BlockType.LOCALS.value, 'bucket_name')
-        self.compare_vertex_attributes(local_graph, expected_resource, BlockType.RESOURCE.value, 'aws_s3_bucket.template_bucket')
+        self.compare_vertex_attributes(local_graph, expected_local, BlockType.LOCALS, 'bucket_name')
+        self.compare_vertex_attributes(local_graph, expected_resource, BlockType.RESOURCE, 'aws_s3_bucket.template_bucket')
 
     def test_render_variable(self):
         resources_dir = os.path.join(TEST_DIRNAME, '../resources/variable_rendering/render_variable')
-        graph_manager = GraphManager('acme', ['acme'])
+        graph_manager = TerraformGraphManager('acme', ['acme'])
         local_graph, _ = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=True)
 
         expected_resource = {'region': "us-west-2", 'bucket': "test_bucket_name", "acl": "acl", "force_destroy": True}
 
-        self.compare_vertex_attributes(local_graph, expected_resource, BlockType.RESOURCE.value, 'aws_s3_bucket.template_bucket')
+        self.compare_vertex_attributes(local_graph, expected_resource, BlockType.RESOURCE, 'aws_s3_bucket.template_bucket')
 
     def test_render_local_from_variable(self):
         resources_dir = os.path.join(TEST_DIRNAME,
                                      '../resources/variable_rendering/render_local_from_variable')
-        graph_manager = GraphManager('acme', ['acme'])
+        graph_manager = TerraformGraphManager('acme', ['acme'])
         local_graph, _ = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=True)
 
         expected_local = {'bucket_name': 'test_bucket_name'}
 
-        self.compare_vertex_attributes(local_graph, expected_local, BlockType.LOCALS.value, 'bucket_name')
+        self.compare_vertex_attributes(local_graph, expected_local, BlockType.LOCALS, 'bucket_name')
 
     def test_general_example(self):
         resources_dir = os.path.join(TEST_DIRNAME, '../resources/general_example')
-        graph_manager = GraphManager('acme', ['acme'])
+        graph_manager = TerraformGraphManager('acme', ['acme'])
         local_graph, _ = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=True)
 
         expected_provider = {'profile': 'default', 'region': 'us-east-1', 'alias': 'east1'}
         expected_local = {'bucket_name': {'val': 'MyBucket'}}
         expected_resource = {'region': 'us-west-2', 'bucket': expected_local['bucket_name']}
 
-        self.compare_vertex_attributes(local_graph, expected_provider, BlockType.PROVIDER.value, 'aws.east1')
-        self.compare_vertex_attributes(local_graph, expected_local, BlockType.LOCALS.value, 'bucket_name')
-        self.compare_vertex_attributes(local_graph, expected_resource, BlockType.RESOURCE.value, 'aws_s3_bucket.template_bucket')
+        self.compare_vertex_attributes(local_graph, expected_provider, BlockType.PROVIDER, 'aws.east1')
+        self.compare_vertex_attributes(local_graph, expected_local, BlockType.LOCALS, 'bucket_name')
+        self.compare_vertex_attributes(local_graph, expected_resource, BlockType.RESOURCE, 'aws_s3_bucket.template_bucket')
 
     def test_terragoat_db_app(self):
         resources_dir = os.path.join(TEST_DIRNAME, '../resources/variable_rendering/render_terragoat_db_app')
-        graph_manager = GraphManager('acme', ['acme'])
+        graph_manager = TerraformGraphManager('acme', ['acme'])
         local_graph, _ = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=True)
 
-        self.compare_vertex_attributes(local_graph, expected_terragoat_local_resource_prefix, BlockType.LOCALS.value, 'resource_prefix')
-        self.compare_vertex_attributes(local_graph, expected_terragoat_db_instance, BlockType.RESOURCE.value, "aws_db_instance.default")
+        self.compare_vertex_attributes(local_graph, expected_terragoat_local_resource_prefix, BlockType.LOCALS, 'resource_prefix')
+        self.compare_vertex_attributes(local_graph, expected_terragoat_db_instance, BlockType.RESOURCE, "aws_db_instance.default")
 
     def test_render_nested_modules(self):
         resources_dir = os.path.join(TEST_DIRNAME, '../resources/variable_rendering/render_nested_modules')
-        graph_manager = GraphManager('acme', ['acme'])
+        graph_manager = TerraformGraphManager('acme', ['acme'])
         local_graph, _ = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=True)
 
         expected_aws_instance = {"instance_type": "bar"}
@@ -78,14 +79,14 @@ class TestRenderer(TestCase):
     def compare_vertex_attributes(self, local_graph, expected_attributes, block_type, block_name):
         vertex = local_graph.vertices[local_graph.vertices_block_name_map[block_type][block_name][0]]
         print(f'breadcrumbs = {vertex.breadcrumbs}')
-        vertex_attributes = vertex.get_decoded_attribute_dict()
+        vertex_attributes = vertex.get_attribute_dict()
         for attribute_key, expected_value in expected_attributes.items():
             actual_value = vertex_attributes.get(attribute_key)
             self.assertEqual(expected_value, actual_value, f'error during comparing {block_type} in attribute key: {attribute_key}')
 
     def test_breadcrumbs(self):
         resources_dir = os.path.join(TEST_DIRNAME, '../resources/s3_bucket')
-        graph_manager = GraphManager('acme', ['acme'])
+        graph_manager = TerraformGraphManager('acme', ['acme'])
         local_graph, _ = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=True)
         vertices = local_graph.vertices
         s3_vertex = list(filter(lambda vertex:  vertex.block_type == BlockType.RESOURCE, vertices))[0]
@@ -105,7 +106,7 @@ class TestRenderer(TestCase):
 
     def test_multiple_breadcrumbs(self):
         resources_dir = os.path.join(TEST_DIRNAME, '../resources/general_example')
-        graph_manager = GraphManager('acme', ['acme'])
+        graph_manager = TerraformGraphManager('acme', ['acme'])
         local_graph, _ = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=True)
         vertices = local_graph.vertices
         s3_vertex = list(filter(lambda vertex:  vertex.block_type == BlockType.RESOURCE, vertices))[0]
@@ -125,16 +126,16 @@ class TestRenderer(TestCase):
 
     def test_render_lambda(self):
         resources_dir = os.path.join(TEST_DIRNAME, '../resources/variable_rendering/render_lambda')
-        graph_manager = GraphManager('acme', ['acme'])
+        graph_manager = TerraformGraphManager('acme', ['acme'])
         local_graph, _ = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=True)
 
         expected_aws_lambda_permission = {'count': 0, 'statement_id': 'test_statement_id', 'action': 'lambda:InvokeFunction', 'function_name': 'my-func', 'principal': 'dumbeldor', 'resource_type': 'aws_lambda_permission'}
 
-        self.compare_vertex_attributes(local_graph, expected_aws_lambda_permission, BlockType.RESOURCE.value, "aws_lambda_permission.test_lambda_permissions")
+        self.compare_vertex_attributes(local_graph, expected_aws_lambda_permission, BlockType.RESOURCE, "aws_lambda_permission.test_lambda_permissions")
 
     def test_eks(self):
         resources_dir = os.path.join(TEST_DIRNAME, '../resources/variable_rendering/terraform-aws-eks-master')
-        graph_manager = GraphManager('eks', ['eks'])
+        graph_manager = TerraformGraphManager('eks', ['eks'])
         local_graph, tf_def = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=True)
 
         for v in local_graph.vertices:
@@ -148,7 +149,7 @@ class TestRenderer(TestCase):
 
     def test_dict_tfvar(self):
         resources_dir = os.path.join(TEST_DIRNAME, '../resources/variable_rendering/render_dictionary_tfvars')
-        graph_manager = GraphManager('d', ['d'])
+        graph_manager = TerraformGraphManager('d', ['d'])
         local_graph, tf_def = graph_manager.build_graph_from_source_directory(resources_dir, render_variables=True)
 
         for v in local_graph.vertices:
@@ -161,7 +162,7 @@ class TestRenderer(TestCase):
 
     def test_graph_rendering_order(self):
         resource_path = os.path.join(TEST_DIRNAME, "..", "resources", "module_rendering", "example")
-        graph_manager = GraphManager('m', ['m'])
+        graph_manager = TerraformGraphManager('m', ['m'])
         local_graph, tf_def = graph_manager.build_graph_from_source_directory(resource_path, render_variables=True)
         module_vertices = list(filter(lambda v: v.block_type == BlockType.MODULE, local_graph.vertices))
         existing = set()
